requirements.txt
Django==3.1.2
djangorestframework==3.12.1
cloudinary==1.22.0
gunicorn==20.0.4 
whitenoise==5.2.0
psycopg2-binary==2.8.6
dj-database-url==0.5.0
django-cors-headers==3.5.0